Fixed test files to match singleuser in components and store
