+++ client/components/SingleUser.js
@@ -1,0 +1,34 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import {connect} from 'react-redux'
+
+/**
+ * COMPONENT
+ */
+export const SingleUser = (props) => {
+  const {email} = props
+
+  return (
+    <div>
+      <h3>Welcome, {email}</h3>
+    </div>
+  )
+}
+
+/**
+ * CONTAINER
+ */
+const mapState = (state) => {
+  return {
+    email: state.user.email
+  }
+}
+
+export default connect(mapState)(SingleUser)
+
+/**
+ * PROP TYPES
+ */
+SingleUser.propTypes = {
+  email: PropTypes.string
+}
